requirements.txt
mysql-connector-python
sqlalchemy
tornado
wtforms
werkzeug